refactor: refactor project struct
