+++ internal/gen/cardinality.go
