--- internal/naming/naming.go
@@ -1,2 +1,0 @@
-// Package naming contains some utilities for generating names.
-package naming
